tagwiseapp/reader/requirements.txt
django>=4.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
webdriver-manager>=3.8.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0 